serverport_ was a SHORT?? huh?
git-svn-id: https://svn.opendtect.org/od/trunk@36163 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Network/netreqcommunic.h
@@ -73,7 +73,7 @@ private:
     TcpConnection*		tcpconn_;
 
     BufferString		servername_;
-    short			serverport_;
+    int				serverport_;
 
     void			connectToHost();
     void			connCloseCB(CallBacker*);
